Fix long instead of int as ID, fix styling

+++ src/PasswordManager.API.Core/Services/IPasswordService.cs
@@ -8,11 +8,11 @@ public interface IPasswordService
 {
     Task<Password> AddPasswordAsync(Password password);
 
-    Task<Password?> GetPasswordAsync(int id);
+    Task<Password?> GetPasswordAsync(long id);
 
     Task<IEnumerable<Password>> GetUserPasswordsAsync();
 
     Task<Password> UpdatePasswordAsync(Password password);
 
-    Task<bool> DeletePasswordAsync(int id);
+    Task<bool> DeletePasswordAsync(long id);
 }
